Create LP from frontend working

+++ smartcontracts/scripts/copyAbis.js
@@ -1,7 +1,16 @@
 const fs = require("fs");
 const path = require("path");
 
-const contracts = ["LPRouter", "EthTokenExchange"]; // Add your contract names here
+const contracts = [
+  "LPRouter",
+  "EthTokenExchange",
+  "IERC20",
+  "ERC20",
+  "LP",
+  "LPRouter",
+  "MinPath",
+  "Queue",
+]; // Add your contract names here
 
 contracts.forEach((contractName) => {
   const artifactPath = path.join(
@@ -12,6 +21,11 @@ contracts.forEach((contractName) => {
     __dirname,
     `../../frontend/src/contracts/abis/${contractName}.json`
   );
+
+  if (!fs.existsSync(artifactPath)) {
+    console.log(`Artifact for ${contractName} not found at ${artifactPath}`);
+    return;
+  }
 
   fs.copyFileSync(artifactPath, destinationPath);
   console.log(
